requirements.txt
aniso8601==7.0.0
bcrypt==3.1.7
cffi==1.13.2
Click==7.0
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-GraphQL==2.0.0
Flask-GraphQL-Auth==1.3.0
flask-mongoengine==0.9.5
Flask-PyMongo==2.3.0
Flask-WTF==0.14.2
graphene==2.1.8
graphene-mongo==0.2.8
graphql-core==2.2.1
graphql-relay==2.0.0
graphql-server-core==1.1.1
iso8601==0.1.12
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
mongoengine==0.18.2
promise==2.2.1
pycparser==2.19
PyJWT==1.7.1
pymongo==3.9.0
Rx==1.6.1
singledispatch==3.4.0.3
six==1.13.0
Werkzeug==0.16.0
WTForms==2.2.1
